backend/requirements.txt
asgiref==3.10.0
binary==1.0.2
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
googlemaps==4.10.0
idna==3.10
msgpack==1.1.2
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
